Factor out instr mixin.
This way I can auto-detect the instruction parser methods to put in the
assembly namespace.

I added in the rest of the asm code for the basic avrvm interpreter and
converted it to use the fancy pythonic format.  Then I added a bunch of
parser methods to make it scan without errors.

Next I can go back and add bit-pattern backends to the second pass module
to fill in the binary output.

Also, db now handles chars (strings of length one.)